requirements.txt
Flask
gunicorn
google-cloud
marshmallow
apispec
python-dateutil
gensim